refactor(api): delegate route logic to services
