+++ packages/application/src/map.ts
@@ -1,0 +1,216 @@
+export function generateMap(size: number, seed: number) {
+  function mulberry32(a: number) {
+    return function() {
+      let t = (a += 0x6d2b79f5)
+      t = Math.imul(t ^ (t >>> 15), t | 1)
+      t ^= t + Math.imul(t ^ (t >>> 7), t | 61)
+      return ((t ^ (t >>> 14)) >>> 0) / 4294967296
+    }
+  }
+  const rng = mulberry32(seed)
+  const map: string[][] = Array.from({ length: size }, () => Array.from({ length: size }, () => 'grass'))
+  const lakes = 1 + Math.floor(rng() * 3)
+  for (let i = 0; i < lakes; i++) {
+    const cx = Math.floor(rng() * size)
+    const cy = Math.floor(rng() * size)
+    const r = 2 + Math.floor(rng() * Math.max(2, size * 0.12))
+    for (let y = Math.max(0, cy - r); y < Math.min(size, cy + r + 1); y++) {
+      for (let x = Math.max(0, cx - r); x < Math.min(size, cx + r + 1); x++) {
+        const dx = x - cx, dy = y - cy
+        if (dx * dx + dy * dy <= r * r * (0.8 + rng() * 0.6)) map[y][x] = 'water'
+      }
+    }
+  }
+  const forests = 1 + Math.floor(rng() * 4)
+  for (let i = 0; i < forests; i++) {
+    const cx = Math.floor(rng() * size)
+    const cy = Math.floor(rng() * size)
+    const r = 2 + Math.floor(rng() * Math.max(2, size * 0.15))
+    for (let y = Math.max(0, cy - r); y < Math.min(size, cy + r + 1); y++) {
+      for (let x = Math.max(0, cx - r); x < Math.min(size, cx + r + 1); x++) {
+        if (map[y][x] !== 'grass') continue
+        const dx = x - cx, dy = y - cy
+        if (dx * dx + dy * dy <= r * r * (0.7 + rng() * 0.5)) map[y][x] = 'forest'
+      }
+    }
+  }
+  const ridges = 1 + Math.floor(rng() * 2)
+  for (let i = 0; i < ridges; i++) {
+    const vertical = rng() < 0.5
+    const start = Math.floor(rng() * size)
+    const len = Math.floor(size * (0.5 + rng() * 0.5))
+    for (let k = 0; k < len; k++) {
+      const t = k / len
+      const x = vertical ? start : Math.floor(t * (size - 1))
+      const y = vertical ? Math.floor(t * (size - 1)) : start
+      if (map[y][x] === 'grass') map[y][x] = 'mountain'
+    }
+  }
+  return map
+}
+
+export function generateMapChunk(chunkX: number, chunkY: number, chunkSize: number, seed: number) {
+  function mulberry32(a: number) {
+    return function() {
+      let t = (a += 0x6d2b79f5)
+      t = Math.imul(t ^ (t >>> 15), t | 1)
+      t ^= t + Math.imul(t ^ (t >>> 7), t | 61)
+      return ((t ^ (t >>> 14)) >>> 0) / 4294967296
+    }
+  }
+  const chunkSeed = seed + chunkX * 1000 + chunkY * 1000000
+  const rng = mulberry32(chunkSeed)
+  const chunk: string[][] = Array.from({ length: chunkSize }, () => Array.from({ length: chunkSize }, () => 'grass'))
+  const biomeRng = mulberry32(seed + Math.abs(chunkX) * 73 + Math.abs(chunkY) * 137)
+  const biomeValue = biomeRng()
+  let biome = 'temperate'
+  if (biomeValue < 0.15) biome = 'desert'
+  else if (biomeValue < 0.3) biome = 'tundra'
+  else if (biomeValue < 0.45) biome = 'swamp'
+  else if (biomeValue < 0.6) biome = 'mountains'
+  else if (biomeValue < 0.75) biome = 'forest'
+  if (biome === 'desert') {
+    for (let y = 0; y < chunkSize; y++) {
+      for (let x = 0; x < chunkSize; x++) {
+        if (rng() < 0.7) chunk[y][x] = 'desert'
+      }
+    }
+  } else if (biome === 'tundra') {
+    for (let y = 0; y < chunkSize; y++) {
+      for (let x = 0; x < chunkSize; x++) {
+        if (rng() < 0.6) chunk[y][x] = 'tundra'
+      }
+    }
+  } else if (biome === 'swamp') {
+    for (let y = 0; y < chunkSize; y++) {
+      for (let x = 0; x < chunkSize; x++) {
+        if (rng() < 0.4) chunk[y][x] = 'swamp'
+      }
+    }
+  }
+  const waterFeatures = biome === 'desert' ? 0 : (biome === 'swamp' ? 3 : 1 + Math.floor(rng() * 2))
+  for (let i = 0; i < waterFeatures; i++) {
+    const cx = Math.floor(rng() * chunkSize)
+    const cy = Math.floor(rng() * chunkSize)
+    const r = 1 + Math.floor(rng() * Math.max(2, chunkSize * 0.15))
+    for (let y = Math.max(0, cy - r); y < Math.min(chunkSize, cy + r + 1); y++) {
+      for (let x = Math.max(0, cx - r); x < Math.min(chunkSize, cx + r + 1); x++) {
+        const dx = x - cx, dy = y - cy
+        const dist = Math.sqrt(dx * dx + dy * dy)
+        if (dist <= r * (0.7 + rng() * 0.4)) {
+          chunk[y][x] = 'water'
+        }
+      }
+    }
+  }
+  if (rng() < 0.3 && biome !== 'desert') {
+    const riverDirection = rng() < 0.5 ? 'horizontal' : 'vertical'
+    const riverPos = Math.floor(rng() * chunkSize)
+    const riverWidth = 1 + Math.floor(rng() * 2)
+    if (riverDirection === 'horizontal') {
+      for (let x = 0; x < chunkSize; x++) {
+        for (let w = 0; w < riverWidth; w++) {
+          const y = Math.min(chunkSize - 1, riverPos + w)
+          if (chunk[y][x] !== 'mountain') chunk[y][x] = 'water'
+        }
+      }
+    } else {
+      for (let y = 0; y < chunkSize; y++) {
+        for (let w = 0; w < riverWidth; w++) {
+          const x = Math.min(chunkSize - 1, riverPos + w)
+          if (chunk[y][x] !== 'mountain') chunk[y][x] = 'water'
+        }
+      }
+    }
+  }
+  const forestDensity = biome === 'forest' ? 0.8 : (biome === 'desert' ? 0.05 : 0.3)
+  if (rng() < forestDensity) {
+    const forests = 1 + Math.floor(rng() * 3)
+    for (let i = 0; i < forests; i++) {
+      const cx = Math.floor(rng() * chunkSize)
+      const cy = Math.floor(rng() * chunkSize)
+      const r = 2 + Math.floor(rng() * Math.max(2, chunkSize * 0.2))
+      for (let y = Math.max(0, cy - r); y < Math.min(chunkSize, cy + r + 1); y++) {
+        for (let x = Math.max(0, cx - r); x < Math.min(chunkSize, cx + r + 1); x++) {
+          if (chunk[y][x] !== 'water' && chunk[y][x] !== 'mountain') {
+            const dx = x - cx, dy = y - cy
+            if (dx * dx + dy * dy <= r * r * (0.6 + rng() * 0.5)) {
+              chunk[y][x] = 'forest'
+            }
+          }
+        }
+      }
+    }
+  }
+  if (biome === 'mountains' || rng() < 0.4) {
+    const mountains = biome === 'mountains' ? 2 + Math.floor(rng() * 3) : 1
+    for (let i = 0; i < mountains; i++) {
+      const cx = Math.floor(rng() * chunkSize)
+      const cy = Math.floor(rng() * chunkSize)
+      const r = 2 + Math.floor(rng() * Math.max(2, chunkSize * 0.18))
+      for (let y = Math.max(0, cy - r); y < Math.min(chunkSize, cy + r + 1); y++) {
+        for (let x = Math.max(0, cx - r); x < Math.min(chunkSize, cx + r + 1); x++) {
+          const dx = x - cx, dy = y - cy
+          const dist = Math.sqrt(dx * dx + dy * dy)
+          if (dist <= r * (0.5 + rng() * 0.4)) {
+            chunk[y][x] = 'mountain'
+          }
+        }
+      }
+    }
+  }
+  const villageChance = biome === 'desert' ? 0.02 : (biome === 'mountains' ? 0.03 : 0.05)
+  if (rng() < villageChance) {
+    let attempts = 0
+    while (attempts < 20) {
+      const vx = 2 + Math.floor(rng() * (chunkSize - 4))
+      const vy = 2 + Math.floor(rng() * (chunkSize - 4))
+      if (chunk[vy][vx] === 'grass' || chunk[vy][vx] === 'forest') {
+        chunk[vy][vx] = 'village'
+        const villageSize = 1 + Math.floor(rng() * 3)
+        for (let dy = -villageSize; dy <= villageSize; dy++) {
+          for (let dx = -villageSize; dx <= villageSize; dx++) {
+            const nx = vx + dx, ny = vy + dy
+            if (nx >= 0 && nx < chunkSize && ny >= 0 && ny < chunkSize) {
+              if (Math.abs(dx) + Math.abs(dy) <= villageSize && rng() < 0.6) {
+                if (chunk[ny][nx] !== 'water' && chunk[ny][nx] !== 'mountain') {
+                  chunk[ny][nx] = dx === 0 && dy === 0 ? 'village' : 'village_outskirt'
+                }
+              }
+            }
+          }
+        }
+        break
+      }
+      attempts++
+    }
+  }
+  if (rng() < 0.08) {
+    let attempts = 0
+    while (attempts < 15) {
+      const rx = Math.floor(rng() * chunkSize)
+      const ry = Math.floor(rng() * chunkSize)
+      if (chunk[ry][rx] === 'grass' || chunk[ry][rx] === 'desert') {
+        chunk[ry][rx] = 'ruins'
+        break
+      }
+      attempts++
+    }
+  }
+  const resourceChance = 0.1
+  if (rng() < resourceChance) {
+    let attempts = 0
+    while (attempts < 20) {
+      const rx = Math.floor(rng() * chunkSize)
+      const ry = Math.floor(rng() * chunkSize)
+      if (chunk[ry][rx] === 'grass' || chunk[ry][rx] === 'forest' || chunk[ry][rx] === 'desert') {
+        const resources = ['iron', 'gold', 'mana_crystal', 'grain_patch']
+        chunk[ry][rx] = resources[Math.floor(rng() * resources.length)]
+        break
+      }
+      attempts++
+    }
+  }
+  return chunk
+}
+
