+++ src/app/api/proposals/[id]/decide/route.ts
@@ -1,7 +1,6 @@
 import { NextRequest, NextResponse } from 'next/server'
-import { createSupabaseServerClient } from '@/lib/supabase/server'
-import { SupabaseUnitOfWork } from '@/infrastructure/supabase/unit-of-work'
 import { z } from 'zod'
+import { decideProposal } from '@application'
 
 interface RouteContext {
   params: Promise<{ id: string }>
@@ -14,8 +13,6 @@ const BodySchema = z.object({
 
 export async function POST(req: NextRequest, context: RouteContext) {
   const params = await context.params
-  const supabase = createSupabaseServerClient()
-  const uow = new SupabaseUnitOfWork(supabase)
   const { id } = params
   const json = await req.json().catch(() => ({}))
   const parsed = BodySchema.safeParse(json)
@@ -26,22 +23,11 @@ export async function POST(req: NextRequest, context: RouteContext) {
     )
   }
   const { decision, comment } = parsed.data
-
-  const prop = await uow.proposals.getById(id)
-  if (!prop) return NextResponse.json({ error: 'Proposal not found' }, { status: 404 })
-
-  // Record decision
-  const { error: decErr } = await supabase.from('decisions').insert({ proposal_id: id, decision, comment })
-  if (decErr) return NextResponse.json({ error: decErr.message }, { status: 500 })
-
-  // Update proposal status
-  const status = decision === 'accept' ? 'accepted' : 'rejected'
   try {
-    await uow.proposals.update(id, { status })
-  } catch (upErr: unknown) {
-    const message = upErr instanceof Error ? upErr.message : String(upErr)
+    await decideProposal(id, decision, comment)
+    return NextResponse.json({ ok: true })
+  } catch (error: unknown) {
+    const message = error instanceof Error ? error.message : String(error)
     return NextResponse.json({ error: message }, { status: 500 })
   }
-
-  return NextResponse.json({ ok: true })
 }
