+++ src/app/api/proposals/route.ts
@@ -1,22 +1,10 @@
-import { NextRequest, NextResponse } from 'next/server'
-import { createSupabaseServerClient } from '@/lib/supabase/server'
-import { SupabaseUnitOfWork } from '@/infrastructure/supabase/unit-of-work'
+import { NextResponse } from 'next/server'
 import logger from '@/lib/logger'
+import { listProposals } from '@application'
 
-export async function GET(req: NextRequest) {
+export async function GET() {
   try {
-    const supabase = createSupabaseServerClient()
-    const uow = new SupabaseUnitOfWork(supabase)
-
-    const state = await uow.gameStates.getLatest()
-    if (!state) return NextResponse.json({ proposals: [] })
-
-    const proposals = await uow.proposals.listByState(state.id, [
-      'pending',
-      'accepted',
-      'rejected',
-    ])
-
+    const proposals = await listProposals()
     return NextResponse.json({ proposals })
   } catch (error) {
     logger.error('Supabase connection error in proposals route:', error)
